# docs/openapi.yaml
openapi: 3.0.0

info:
  title: Blog API
  version: 1.0.0
  description: API REST para gestión de authors y posts

servers:
  - url: http://localhost:3000

tags:
  - name: Authors
    description: Operaciones sobre autores
  - name: Posts
    description: Operaciones sobre publicaciones

paths:

  /authors:
    get:
      tags: [Authors]
      summary: Obtener todos los authors
      operationId: getAllAuthors
      responses:
        '200':
          description: Lista de authors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorList'

    post:
      tags: [Authors]
      summary: Crear un author
      operationId: createAuthor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorInput'
            example:
              name: Juan
              email: juan@test.com
              bio: Autor
      responses:
        '201':
          description: Author creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /authors/{id}:
    get:
      tags: [Authors]
      summary: Obtener author por ID
      operationId: getAuthorById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Author encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '404':
          description: Author no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Authors]
      summary: Actualizar author
      operationId: updateAuthor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorInput'
      responses:
        '200':
          description: Author actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Author no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Authors]
      summary: Eliminar author
      operationId: deleteAuthor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado
        '404':
          description: Author no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /posts:
    get:
      tags: [Posts]
      summary: Obtener todos los posts
      operationId: getAllPosts
      responses:
        '200':
          description: Lista de posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostList'

    post:
      tags: [Posts]
      summary: Crear post
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
            example:
              author_id: 1
              title: Mi post
              content: Contenido
              published: true
      responses:
        '201':
          description: Post creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /posts/{id}:
    get:
      tags: [Posts]
      summary: Obtener post por ID
      operationId: getPostById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Posts]
      summary: Actualizar post
      operationId: updatePost
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '200':
          description: Post actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Posts]
      summary: Eliminar post
      operationId: deletePost
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /posts/author/{authorId}:
    get:
      tags: [Posts]
      summary: Obtener posts por author
      operationId: getPostsByAuthor
      parameters:
        - in: path
          name: authorId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de posts del author
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostList'

components:
  schemas:

    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Error de validación

    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: Operación realizada correctamente

    Author:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        email:
          type: string
        bio:
          type: string
        created_at:
          type: string
          format: date-time

    AuthorInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        bio:
          type: string

    AuthorResponse:
      type: object
      properties:
        message:
          type: string
          example: Author creado correctamente
        data:
          $ref: '#/components/schemas/Author'

    AuthorList:
      type: array
      items:
        $ref: '#/components/schemas/Author'

    Post:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        author_id:
          type: integer
        title:
          type: string
        content:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time

    PostInput:
      type: object
      required: [author_id, title, content]
      properties:
        author_id:
          type: integer
        title:
          type: string
        content:
          type: string
        published:
          type: boolean

    PostResponse:
      type: object
      properties:
        message:
          type: string
          example: Post creado correctamente
        data:
          $ref: '#/components/schemas/Post'

    PostList:
      type: array
      items:
        $ref: '#/components/schemas/Post'